changing the wrappers to pure functions, removing fetch logic
