+++ packages/calendar/__fixtures__/calendar.js
@@ -1,11 +1,12 @@
 import Calendar from '..';
 // eslint-disable-next-line import/extensions
 import events from './events.json';
+import elvantoEvents from '../elvanto-wrapper';
 
 export default {
   component: Calendar,
   props: {
-    events,
+    events: elvantoEvents(events),
     initialView: 'month'
   }
 };
